refactor: rename API classes to improve package structure and organization
