+++ src/main/java/org/thingai/app/api/endpoints/MatchControlApi.java
@@ -1,10 +1,11 @@
-package org.thingai.app.api.v2;
+package org.thingai.app.api.endpoints;
 
 import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RestController;
 
 @RestController
-@RequestMapping("/api/rank")
-public class RankApi {
+@RequestMapping("/api/scorekeeper")
+public class MatchControlApi {
+
 
 }
